extract mailchimp updates as a concern which can be applied to champions
