+++ app/models/user.rb
@@ -40,6 +40,8 @@ end
 # With that fixed, we can now define the regular User class.
 
 class User < ActiveRecord::Base
+  include MailchimpUpdates
+  
   # Include default devise modules. Others available are:
   # :confirmable, :lockable, :timeoutable and :omniauthable
   devise :registerable,
@@ -57,32 +59,10 @@ class User < ActiveRecord::Base
   validates :last_name, presence: true
 
   before_save :capitalize_name
-  before_save :update_mailchimp
   
   def capitalize_name
     self.first_name = first_name.split.map(&:capitalize).join(' ') unless first_name.nil?
     self.last_name = last_name.split.map(&:capitalize).join(' ') unless last_name.nil?
-  end
-  
-  def update_mailchimp
-    # don't update for a new record, it doesn't exist in mailchimp yet
-    return true if new_record?
-
-    mailchimp = BeyondZ::Mailchimp::User.new(self)
-    
-    # don't update unless an updateable field has changed
-    return true unless mailchimp.requires_update?
-    
-    success_status = mailchimp.update
-
-    # for now, assume success until retro-sync can be performed
-    success_status = true
-    
-    unless success_status
-      self.errors[:email] << "could not be updated on MailChimp"
-    end
-    
-    success_status
   end
 
   # Finds the lead owner from the uploaded spreadsheet mapping, or returns
